Local player prediction
Using a std::deque of input messages, a server snapshot is applied,
and those local inputs that have not yet been seen by the server are
replayed, the rest is discarded from the deque.

The interpolation function is just a dummy for now,
it's needed to update the enemy players, else they don't move at all.

Tested on localhost @ 60FPS/5Hz.
There's occasional desync indicated by "Reconciliation error" warnings, by not moving for a bit, it resyncs.
This gets worse on higher server tickrates though

Bugfix: Inflight inputs

Previously the client did not distinguish between local intputs and inputs that are in-flight (on the way to the server),
thus, on reconciliation some inputs were inadvertently applied mutliple times, which lead to sporadic bursts of reconciliation errors (manifesting as teleport hops).

The solution is to store and accumulate local inputs that are not yet in-flight separately from in-flight inputs.

Bugfix: Maintain fixed tick rate updates

Previously, the server would wait on each matchLoop() iteration until UNRELIABLE_TICK_MS have passed before executing the next tick of simulation+snapshot broadcast.

The problem was that the measurement was restarted before waiting, so the server actually processes two ticks before waiting again.

This caused the local interpolation algorithm to fail, resulting in enemies teleporting back and forth.

The fix is to restart the clock *after* waiting for  UNRELIABLE_TICK_MS to pass.