Bugfix: Clean state for new game
Previously, the GameServer was reused from the last Game when a new Game starts. (The GameClient is already re-initialized every time a new game starts.) This discrepancy causes dirty state on the Server to be carried over into the new game, e.g. last received client ticks.

This causes clients to be unable to move until they reached the tick number when the old game ended, as the server will reject "outdated" packets (outdated as the counter was carried over from the last game).

The fix is to move the GameServer variable into the loop rather then create it once per Server process lifetime.
Like this the object is intitialized every time a new Game starts.