Bugfix: Wall destruction not synced to Client
Right now, the server keeps track of each wall cube's health value, and adds destroyed wall tiles (RawLayer indices) into the WorldState snapshot when a wall gets destroyed.

The client is not informed of a wall's health value, it stays at full health for the entire game.

The client will apply destroyed walls from a snapshot, but only erases RawLayer data (texture render model disappearing), *the logical wall stays untouched*.

Thus, on client side prediction, the client thinks the player is colliding with a wall, simulates a collision with health reduction, only the server knows the wall was destroyed. What follows is a prediction error for every frame the client is moving "inside" the destroyed wall.

This fix ignores the logical m_wall health value and instead checks whether the RawLayer wall has been cleared (indicating it was destroyed).

The current model has deeper issues, there is a disconnection between logical and render wall state. It is not straightforward to translate between these, but they need to be kept in sync.